Code review: addressing small comments
